+++ src/api/index.ts
@@ -4,7 +4,7 @@ export const client = axios.create({
   baseURL: "https://omeda.city",
   headers: {
     "Content-Type": "application/json",
-    "x-origin-sdk": "omedacity-js@0.5.2", // Todo: Automate the incrementation
+    "x-origin-sdk": "omedacity-js@0.6.0", // Todo: Automate the incrementation
   },
 });
 
@@ -22,6 +22,11 @@ export const createQueryParams = <T>(params?: T): string => {
 
     if (typeof param !== "string" && typeof param !== "number") {
       for (const pKey in param) {
+        // TODO: There should be a better way to handle params set to false.
+        if (param[pKey] === false) {
+          continue;
+        }
+
         if (query.length === 0) {
           query.push("?" + key + "[" + pKey + "]" + "=" + param[pKey]);
         } else {