+++ src/collections/builds.ts
@@ -1,4 +1,4 @@
-import { Builds, Build } from "../types";
+import { Builds, Build, BuildsQueryParams } from "../types";
 import { buildsFindAll, buildsFindById } from "../api/builds";
 import { client } from "../api";
 
@@ -8,18 +8,19 @@ const findBuildByIdWithClient = buildsFindById(client);
 export class BuildsCollection {
   /**
    * Get all Builds
+   * @param {BuildsQueryParams} params
    * @returns {Promise<Builds>}
    */
-  get(): Promise<Builds> {
-    return findBuildsWithClient();
+  get(params?: BuildsQueryParams): Promise<Builds> {
+    return findBuildsWithClient(params);
   }
 
   /**
    * Get a Build by BuildId
-   * @param {string} buildId
+   * @param {number} buildId
    * @returns {Promise<Build>}
    */
-  getById(buildId: string): Promise<Build> {
+  getById(buildId: number): Promise<Build> {
     return findBuildByIdWithClient(buildId);
   }
 }